:truck: move LoggingConfigurator to own module
